fix(ruby.gems): do not install or update from Gemfile due to lack of permissions at 42

+++ .config/yadm/bootstrap.d/02-install-software.sh
@@ -14,7 +14,7 @@ fi
 
 gemfile=$HOME/.config/yadm/scripts/Gemfile
 
-if [ -f "$gemfile" ]; then
+if [ -f "$gemfile" ] && [ "$(yadm config local.class)" != "42" ]; then
   echo "Installing Ruby gems..."
   bundle install --gemfile="$gemfile"
 fi
